backend/app/requirements.txt
exceptiongroup==1.2.2
fastapi==0.111.1
httptools==0.6.1
pip==23.1.2
python-dotenv==1.0.1
PyYAML==6.0.1
schema==0.7.7
sse-starlette==2.1.2
tensorflow
watchfiles==0.22.0
websockets==12.0
sqlalchemy
passlib
bcrypt
uvicorn
pyjwt
python-jose
python-dotenv
alembic
pytz
scikit-learn
statsmodels